+++ db/migrate/20101219153505_create_transactions.rb
@@ -3,7 +3,7 @@ class CreateTransactions < ActiveRecord::Migration
     create_table :transactions do |t|
       t.integer :ref
       t.text :desc
-      t.integer :value
+      t.float :value
       t.timestamps
     end
   end
